fixed deletion of ckpt dir

+++ run_experiment.py
@@ -11,7 +11,8 @@ if __name__ == "__main__":
     config = Config.from_yaml()
 
     # check_for_checkpoints(config.paths.checkpoint_dir)
-    shutil.rmtree(config.paths.checkpoint_dir)
+    if os.path.exists(config.paths.checkpoint_dir):
+        shutil.rmtree(config.paths.checkpoint_dir)
 
     trainer = latentModulatedTrainer(config.model, config.training, config.paths).to(
         config.model.device
